feat: SafeSmartAccount::transaction_transfer fix + E2E tests (#72)
