+++ bedrock/src/primitives/tooling_tests.rs
@@ -284,7 +284,7 @@ impl FileSystemTester {
         content: &str,
     ) -> Result<(), FileSystemTestError> {
         // _bedrock_fs is automatically injected by the macro
-        // FileSystemError automatically converts to FileSystemTestError::FileSystem
+        // `FileSystemError` automatically converts to FileSystemTestError::FileSystem
         Ok(_bedrock_fs.write_file(filename, content.as_bytes().to_vec())?)
     }
 
@@ -296,7 +296,7 @@ impl FileSystemTester {
         &self,
         filename: &str,
     ) -> Result<String, FileSystemTestError> {
-        // FileSystemError from _bedrock_fs automatically converts to FileSystemTestError::FileSystem
+        // `FileSystemError` from _bedrock_fs automatically converts to FileSystemTestError::FileSystem
         let data = _bedrock_fs.read_file(filename)?;
         String::from_utf8(data).map_err(|_| FileSystemTestError::TestError {
             message: "Invalid UTF-8 data".to_string(),
